+++ items/_items.mjs
@@ -2,6 +2,7 @@ import gratefulFeyCharm from "./gratefulFeyCharm.mjs";
 import guardianScales from "./guardianScales.mjs"
 import mantleOfTheArcaneTrickster from "./mantleOfTheArcaneTrickster.mjs";
 import viceGrip from "./viceGrip.mjs";
+import wyrmreaverGauntlets from "./wyrmreaverGauntlets.mjs";
 
 export default {
     registerSection() {
@@ -9,6 +10,7 @@ export default {
         guardianScales.register();
         mantleOfTheArcaneTrickster.register();
         viceGrip.register();
+        wyrmreaverGauntlets.register();
     }
 }